Bump file version to start in Login phase
Also deprecate lots of old, unused functionality.

Note that API-wise this change isn't breaking.
As such, you need to pass an additional parameter to the ReplayInputStream constructor to indicate your willingness to receive login phase packets.

This is required for the 1.13 Forge handshake.
Neither encryption nor compression packets must be included.
If there's no significant handshake in the login phase (e.g. vanilla), just the LoginSuccessPacket at the beginning of the replay stream is sufficient (as is done in this commit for older versions).