Fix map data being swallowed by squash filter
The MapData packet may contain partial updates, so we cannot just keep the
latest one because it may not contain all required data.
We could parse the updates and track the full state like we do for chunks but
that's probably not worth the effort, so we just put all of them into the
world-specific list.