Fix incorrect packet release during SquashFilter.flush
When onEnd is called, the ownership over all packets is transferred to the
provided list. But the internal lists were not consistently cleared, so the
subsequent release call actually freed most of the packets even though we still
need them.

Technically it should be sufficient to just remove the release call but this
commit additionally moves the resetting of the internal lists into onEnd where
it conceptually belongs, so we do not (even temporarily) hold on to packets we
no longer own.