Fix OOB when converting local palette storage to global palette
That is because the `this.bitsPerEntry` value has already been increased by this
point but we use it to determine how to access the old states.
We can actually just remove that check cause we will never resize again once we
have switched to the global palette, so that case will never happen.