Rewrite JoinGame packet (fixes ReplayMod/ReplayMod#386)
Minecraft 1.8.9 (vanilla, not Forge) handles the entity armor packet differently
depending on the entity class. In particular, all the slots are offset for the
client player entity compared to other player entities. ReplayMod however
injects the client player entity into a replay as if it was just another player.
This doesn't cause any issues during regular playback because the recording
player entity is truly just a regular server player. It does cause issues though
when the replay is played via ViaVersion because it determines the client player
id by looking at the join packet, which by default has the same id as the fake
client player. And therefore ViaVersion re-writes our equipment packets as if
they were for a client player entity, breaking where the armor ends up in newer
versions.
To fix this behavior, we now re-write the entity id in the JoinPacket to match
the camera entity, right after reading it from the stream before we pass it on
to ViaVersion. ViaVersion will then read the camera entity id as the client
player and no longer break equipment packets destined for the fake client
player.